+++ renku_notebooks/errors/intermittent.py
@@ -53,9 +53,9 @@ class CannotStartServerError(IntermittentError):
 class JSCacheError(IntermittentError):
     """Raised when the jupyter server cache responds with anything other than a 200 status
     code. This indicates trouble with the path requested (i.e. the jupyter cache is not aware
-    of the path) or the jypyter server cache is not functioning properly. When this error
+    of the path) or the jupyter server cache is not functioning properly. When this error
     is raised the regular (non-cached) k8s client takes over the fulfils the request. Please
-    note that this is possible because the jypyter server cache will respond with 200 and
+    note that this is possible because the jupyter server cache will respond with 200 and
     an empty response if resource that do not exist are requested."""
 
     message: str = "The jupyter server cache produced and unexpected error."
@@ -63,8 +63,33 @@ class JSCacheError(IntermittentError):
 
 
 class RetryTimeoutError(IntermittentError):
-    """Raised when something was expected to be retried and to evenentually succeed but was retried
+    """Raised when something was expected to be retried and to eventually succeed but was retried
     too many times so that it timed out."""
 
     message: str = "Retrying the request timed out."
     code: int = IntermittentError.code + 4
+
+
+@dataclass
+class PatchServerError(IntermittentError):
+    """Raised when a user server cannot be patched."""
+
+    message: str = "The server cannot be patched."
+    code: int = IntermittentError.code + 5
+
+
+@dataclass
+class PVDisabledError(IntermittentError):
+    """Raised when cannot hibernating because PVs aren't enabled in the config."""
+
+    message: str = "Persistent Volumes aren't enabled in the config."
+    code: int = IntermittentError.code + 6
+
+
+@dataclass
+class AnonymousUserPatchError(IntermittentError):
+    """Raised when trying to patch an anonymous user's session."""
+
+    message: str = "Cannot patch sessions of anonymous users."
+    code: int = IntermittentError.code + 7
+    status_code: int = 422